renaming coding to codingClient
